Restructure some utilities into the training module
